+++ commands/game/protect.js
@@ -5,7 +5,7 @@ module.exports = {
     name: "protect",
     description: "Protect a player from being attacked.",
     usage: `${process.env.PREFIX}protect <player>`,
-    aliases: ["heal", "save"],
+    aliases: ["heal", "save", "antidote"],
     gameOnly: true,
     run: async (message, args, client) => {
         const gamePhase = db.get(`gamePhase`)
@@ -15,11 +15,12 @@ module.exports = {
         if (!message.channel.name.startsWith("priv")) return // if they are not in the private channel
 
         if (player.status !== "Alive") return await message.channel.send("Listen to me, you need to be ALIVE to protect players.")
-        if (!["Doctor", "Bodyguard", "Tough Guy", "Witch", "Ghost Lady", "Night Watchman"].includes(player.role) && !["Doctor", "Bodyguard", "Tough Guy", "Witch", "Ghost Lady", "Night Watchman"].includes(player.dreamRole)) return
-        if (["Doctor", "Bodyguard", "Tough Guy", "Witch", "Ghost Lady", "Night Watchman"].includes(player.dreamRole)) player = db.get(`player_${player.target}`)
+        if (!["Doctor", "Bodyguard", "Tough Guy", "Witch", "Ghost Lady", "Night Watchman", "Lethal Seer"].includes(player.role) && !["Doctor", "Bodyguard", "Tough Guy", "Witch", "Ghost Lady", "Night Watchman", "Lethal Seer"].includes(player.dreamRole)) return
+        if (["Doctor", "Bodyguard", "Tough Guy", "Witch", "Ghost Lady", "Night Watchman", "Lethal Seer"].includes(player.dreamRole)) player = db.get(`player_${player.target}`)
         if (gamePhase % 3 != 0) return await message.channel.send("You do know that you can only protect during the night right? Or are you delusional?")
         if (player.jailed) return await message.channel.send("You are jailed. You cannot use your abilities while in jail!")
         if (player.nightmared) return await message.channel.send("You are nightmared. You cannot use your abilities while you're asleep.")
+        if (player.role !== "Lethal Seer" && args.length !== 1) return await message.channel.send("Please select a player to protect!")
         if (["Night Watchman", "Ghost Lady"].includes(player.role) && player.uses === 0) return await message.channel.send("You have already used up your abilities!")
 
         let object = {
@@ -29,11 +30,17 @@ module.exports = {
             Witch: getEmoji("potion", client),
             "Ghost Lady": getEmoji("gl_protection", client),
             "Night Watchman": getEmoji("nwm_select", client),
+            "Lethal Seer": getEmoji("lethal_seer", client)
         }
 
-        if (args[0].toLowerCase() === "cancel") {
+        if (args[0].toLowerCase() === "cancel" && player.role !== "Lethal Seer") {
             db.delete(`player_${player.id}.target`)
             return await message.channel.send(`${object[player.role]} Done! That player is no longer protected!`)
+        }
+
+        if (player.role === "Lethal Seer") {
+            db.delete(`player_${player.id}.target`)
+            return await message.channel.send(`${object[player.role]} You have used your antidote so the player you checked will not be killed tonight!`)
         }
 
         let target = players[Number(args[0]) - 1] || players.find((p) => p === args[0]) || players.map((p) => db.get(`player_${p}`)).find((p) => p.username === args[0])
